requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
yfinance>=0.2.35
requests>=2.31.0
scipy>=1.11.0
finnhub-python>=2.4.19
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
